adminservice/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
python-dotenv
pydantic
databases
asyncpg
psycopg2-binary